fix(coding-agent): correct todo widget windowing
