+++ packages/coding-agent/src/core/extensions/builtin/todotools/changes.md
@@ -7,14 +7,22 @@
 - Bounded the phase-aware `todo-sidebar` output to the interactive widget's
   10-line budget before the generic widget renderer truncates it.
 - Long active phases now keep the two tasks immediately before the active
-  item, the active item itself, and as much upcoming work as fits.
-- Explicit earlier/later omission rows report how many tasks are outside the
-  window. Short phases preserve their existing complete output.
+  item, the active item itself, and as much pending upcoming work as fits.
+- Completed and abandoned tasks after the active item no longer consume the
+  upcoming-work window or inflate the later omission count.
+- Explicit earlier/later omission rows report how much relevant work is
+  outside the window. Short phases preserve their existing complete output.
+- Split the former monolithic `state.ts` implementation into focused state,
+  query, resolution, operation, formatting, storage, and widget modules, each
+  below the 250-pure-line ceiling. Widget regressions now live in a dedicated
+  deterministic test file.
 
 ### Expected merge conflict zones
 
-- MEDIUM: `state.ts` around `getTodoWidgetLines` if upstream changes todo
-  sidebar layout or the interactive widget line budget.
+- MEDIUM: `todo-widget.ts` if upstream changes todo sidebar layout or the
+  interactive widget line budget.
+- MEDIUM: `state.ts` and the focused `todo-*.ts` modules if upstream changes
+  the todotools state API or persistence behavior.
 
 ## 2026-07-28 - Preserve open todo work across new instructions
