+++ packages/coding-agent/test/suite/todo-extension.test.ts
@@ -556,85 +556,6 @@ describe("todo extension", () => {
 		).toBeUndefined();
 	});
 
-	it("centers a long sidebar window on recent progress and upcoming work", () => {
-		const phases: TodoPhase[] = [
-			{
-				name: "Review",
-				tasks: [
-					{ content: "Completed 1", status: "completed" },
-					{ content: "Completed 2", status: "completed" },
-					{ content: "Completed 3", status: "completed" },
-					{ content: "Completed 4", status: "completed" },
-					{ content: "Completed 5", status: "completed" },
-					{ content: "Completed 6", status: "completed" },
-					{ content: "Current task", status: "in_progress" },
-					{ content: "Upcoming 1", status: "pending" },
-					{ content: "Upcoming 2", status: "pending" },
-					{ content: "Upcoming 3", status: "pending" },
-					{ content: "Upcoming 4", status: "pending" },
-					{ content: "Upcoming 5", status: "pending" },
-				],
-			},
-		];
-
-		expect(getTodoWidgetLines(phases)).toEqual([
-			"Todo",
-			"Review",
-			"... (4 earlier tasks)",
-			"[✓] Completed 5",
-			"[✓] Completed 6",
-			"[•] Current task",
-			"[ ] Upcoming 1",
-			"[ ] Upcoming 2",
-			"[ ] Upcoming 3",
-			"... (2 later tasks)",
-		]);
-	});
-
-	it("uses the active phase as the bounded sidebar window", () => {
-		const phases: TodoPhase[] = [
-			{
-				name: "Closed phase",
-				tasks: [
-					{ content: "Closed 1", status: "completed" },
-					{ content: "Closed 2", status: "completed" },
-				],
-			},
-			{
-				name: "Delivery",
-				tasks: [
-					{ content: "Delivery done 1", status: "completed" },
-					{ content: "Delivery done 2", status: "completed" },
-					{ content: "Delivery done 3", status: "completed" },
-					{ content: "Ship now", status: "in_progress" },
-					{ content: "Verify 1", status: "pending" },
-					{ content: "Verify 2", status: "pending" },
-					{ content: "Verify 3", status: "pending" },
-					{ content: "Verify 4", status: "pending" },
-					{ content: "Verify 5", status: "pending" },
-					{ content: "Verify 6", status: "pending" },
-				],
-			},
-			{ name: "Later phase", tasks: [{ content: "Publish", status: "pending" }] },
-		];
-
-		const lines = getTodoWidgetLines(phases);
-
-		expect(lines).toEqual([
-			"Todo",
-			"Delivery",
-			"... (1 earlier task)",
-			"[✓] Delivery done 2",
-			"[✓] Delivery done 3",
-			"[•] Ship now",
-			"[ ] Verify 1",
-			"[ ] Verify 2",
-			"[ ] Verify 3",
-			"... (3 later tasks)",
-		]);
-		expect(lines).toHaveLength(10);
-	});
-
 	it("renders compact calls and a static phase tree with collapsed closed phases", async () => {
 		const tool = await captureTodoTool();
 		if (!tool.renderCall || !tool.renderResult) throw new Error("Expected todo renderers");
